requirements.txt
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
PyPDF2==3.0.1
python-dotenv==1.0.0
openai==1.3.0
google-generativeai==0.3.0
reportlab==4.0.4